libsite/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.4
django-admin-tools==0.9.1
django-cacheops==6.1
django-debug-toolbar==3.2.2
django-phonenumber-field==6.1.0
django-redis==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
funcy==1.17
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
wrapt==1.14.1
